requirements.txt
PyPDF2
python-dotenv
langchain
langchain-google-genai
